fix sync issue, move embeds to difference src file
