--- .env
@@ -1,11 +1,0 @@
-DISCORD_TOKEN=
-
-BOT_OWNER=""
-
-DISCORD_GUILD=""
-DROPS_CHANNEL=""
-GENERAL_CHANNEL=""
-
-DISCORD_DEV_GUILD=""
-DEV_DROPS_CHANNEL=""
-DEV_GENERAL_CHANNEL=""
